NO-TICKET: Added getShardId to Sharded PGModelInterface
