+++ src/Zynga/Framework/PgData/V1/Interfaces/Sharded/PgModelInterface.hh
@@ -7,4 +7,7 @@ use
 ;
 use Zynga\Framework\Type\V1\Interfaces\TypeInterface;
 
-interface PgModelInterface extends PgModelInterfaceBase {}
+interface PgModelInterface extends PgModelInterfaceBase {
+  public function getShardId(): TypeInterface;
+
+}
